requirements.txt
# requirements.txt - Recipe Extractor API Dependencies
openai>=1.21.0
fastapi
uvicorn[standard]
yt-dlp
pysrt
python-dotenv